requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
sentence-transformers
pytesseract
pillow
requests
python-dotenv
tavily-python